+++ CHANGELOG.md
@@ -7,6 +7,9 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+ -  Tests for XSF writing
+
 ### Fixed
  -  More outstanding bugs for XSF writing